move authOptions to separate file
